# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Contoh API documentation",
        "description": "Contoh API doc menggunakan OpenAPI/Swagger",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "choirudin.emchagmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "paths": {
        "/api/gallery": {
            "get": {
                "tags": [
                    "gallery"
                ],
                "summary": "Tampilkan Gallery",
                "description": "Ini adalah dokumentasi untuk menampilkan gallery",
                "operationId": "gallery_index",
                "responses": {
                    "default": {
                        "description": "Proses Berhasil"
                    }
                }
            }
        },
        "/api/gallery/store": {
            "post": {
                "tags": [
                    "gallery"
                ],
                "summary": "Tambah Gambar",
                "description": "Ini adalah dokumentasi untuk menambah gambar pada gallery",
                "operationId": "gallery.store",
                "requestBody": {
                    "description": "Data untuk mengunggah gambar",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Judul Upload",
                                        "type": "string"
                                    },
                                    "content": {
                                        "description": "Deskripsi Gambar",
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "File Gambar",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "Proses Berhasil"
                    }
                }
            }
        }
    }
}